+++ docs/README.md
@@ -1,4 +1,4 @@
-# Zea User Guide
+# Zea.Zea User Guide
 
 // Update the title above to match the actual product name
