--- src/main/java/ListCommand.java
@@ -1,8 +1,0 @@
-import java.util.ArrayList;
-
-public class ListCommand extends Command {
-    @Override
-    public void execute(Tasks tasks, Ui ui, Storage storage) {
-        ui.displayWithNumbering("Here are your tasks", tasks.tasks);
-    }
-}
