+++ src/main/java/Zea/ParseException.java
@@ -1,3 +1,5 @@
+package Zea;
+
 public class ParseException extends Exception {
 
     public ParseException(String s) {
