+++ src/main/java/Zea/ZeaException.java
@@ -1,3 +1,5 @@
+package Zea;
+
 public class ZeaException extends Exception{
    public ZeaException(String message){
        super(message);
