+++ src/main/java/Zea/task/Deadline.java
@@ -1,3 +1,5 @@
+package Zea.task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
