+++ src/main/java/Zea/task/Event.java
@@ -1,7 +1,9 @@
+package Zea.task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
-public class Event extends Task{
+public class Event extends Task {
 
     protected LocalDateTime to;
     protected LocalDateTime from;
