+++ src/main/java/Zea/task/Task.java
@@ -1,3 +1,5 @@
+package Zea.task;
+
 import java.time.format.DateTimeFormatter;
 
 public abstract class Task {
