+++ src/main/java/Zea/task/Tasks.java
@@ -1,3 +1,5 @@
+package Zea.task;
+
 import java.util.ArrayList;
 
 public class Tasks {
