+++ src/main/java/Zea/task/Todo.java
@@ -1,4 +1,6 @@
-public class Todo extends Task{
+package Zea.task;
+
+public class Todo extends Task {
 
     public Todo(String description) {
         super(description);
