+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 ______________________
-Hello! I'm Zea
+Hello! I'm Zea.Zea
 What can I do for you?
 ______________________
